LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-button-hover-remove-expected.txt


FAIL CSS Test: ::scroll-button changes on hover via :hover and removes correctly assert_equals: ::scroll-button changes on hover via :hover expected "rgb(0, 0, 255)" but got ""
